Move Markdown-preview plugin to 'extensions' folder
