+++ extensions/oni-plugin-markdown-preview/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "oni-markdown-preview",
+  "name": "oni-plugin-markdown-preview",
   "version": "0.0.1",
   "main": "lib/index.js",
   "engines": {
@@ -20,22 +20,18 @@
       }
     }
   },
-  "dependencies": {},
   "tbd-dependencies": {
     "marked": "^0.3.6",
-    "dompurify": "1.0.2"
-  },
-  "tbd-peerDependencies": {
-    "typescript": "^2.5.3",
-    "rimraf": "^2.6.2",
+    "dompurify": "1.0.2",
     "oni-types": "^0.0.4",
     "oni-api": "^0.0.9"
   },
+  "tbd-peerDependencies": {
+    "typescript": "^2.5.3",
+    "rimraf": "^2.6.2"
+  },
   "tbd-devDependencies": {
     "typescript": "2.5.3",
-    "rimraf": "2.6.2",
-    "vscode-languageserver-types": "3.5.0",
-    "oni-types": "0.0.4",
-    "oni-api": "file:../../../node_modules/oni-api"
+    "rimraf": "2.6.2"
   }
 }
